caf: net_state: Move net state to CAF
This commit moves net_state module from pelion_client sample into CAF.
It also moves factory_reset module used by net_state_ot.
