+++ applications/pelion_client/configuration/nrf9160dk_nrf9160ns@1.0.0/app_ZDebug.conf
@@ -9,8 +9,6 @@ CONFIG_PELION_CLIENT=y
 CONFIG_PELION_NRF_SECURITY=y
 CONFIG_PELION_UPDATE_DEVELOPER=y
 
-CONFIG_PELION_CLIENT_FACTORY_RESET_REQUEST_ENABLE=y
-
 ################################################################################
 
 CONFIG_FPU=y
@@ -21,8 +19,12 @@ CONFIG_CAF=y
 CONFIG_CAF_BUTTONS=y
 CONFIG_CAF_BUTTONS_POLARITY_INVERSED=y
 
+CONFIG_CAF_FACTORY_RESET_REQUEST=y
+
 CONFIG_CAF_LEDS=y
 CONFIG_CAF_LEDS_PM_EVENTS=y
+
+CONFIG_CAF_NET_STATE=y
 
 CONFIG_CAF_POWER_MANAGER=y
 CONFIG_CAF_POWER_MANAGER_TIMEOUT=300
