+++ applications/pelion_client/src/events/Kconfig
@@ -8,11 +8,6 @@ menu "Event options"
 
 comment "Initially logged events"
 
-config PELION_INIT_LOG_NET_STATE_EVENTS
-	bool "Log net state events"
-	depends on LOG
-	default y
-
 config PELION_INIT_LOG_PELION_STATE_EVENTS
 	bool "Log Pelion state events"
 	depends on LOG
@@ -22,7 +17,5 @@ config PELION_INIT_LOG_PELION_CREATE_OBJECTS_EVENTS
 	bool "Log Pelion create objects events"
 	depends on LOG
 	default y
-
-rsource "Kconfig.factory_reset_event"
 
 endmenu
