+++ applications/pelion_client/src/modules/CMakeLists.txt
@@ -19,14 +19,5 @@ target_sources_ifdef(CONFIG_BOOTLOADER_MCUBOOT
 target_sources_ifdef(CONFIG_CAF_LEDS
 		     app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/led_state.c)
 
-target_sources_ifdef(CONFIG_LTE_LINK_CONTROL
-		     app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/net_lte.c)
-
-target_sources_ifdef(CONFIG_NET_L2_OPENTHREAD
-		     app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/net_ot.c)
-
-target_sources_ifdef(CONFIG_PELION_CLIENT_FACTORY_RESET_REQUEST_ENABLE
-		     app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/factory_reset_request.c)
-
 target_sources_ifdef(CONFIG_PELION_DEBUG_TRACE
 		     app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/pelion_trace.c)
