+++ applications/pelion_client/src/modules/Kconfig
@@ -5,7 +5,6 @@
 #
 
 rsource "Kconfig.dfu"
-rsource "Kconfig.factory_reset_request"
 rsource "Kconfig.led_state"
 rsource "Kconfig.net"
 rsource "Kconfig.oma_objects"
