+++ include/caf/events/factory_reset_event.h
