+++ subsys/caf/events/Kconfig
@@ -129,7 +129,20 @@ config CAF_INIT_LOG_SENSOR_STATE_EVENTS
 	help
 	  Log sensor state events, used to notify about sensor state change.
 
+config CAF_NET_STATE_EVENTS
+	bool "Enable network state events"
+	help
+	  Enable support for network state events.
 
+config CAF_INIT_LOG_NET_STATE_EVENTS
+	bool "Log network state events"
+	depends on CAF_NET_STATE_EVENTS
+	depends on LOG
+	default y
+	help
+	  Log network state events.
+
+rsource "Kconfig.factory_reset_event"
 rsource "Kconfig.force_power_down_event"
 rsource "Kconfig.keep_alive_event"
 rsource "Kconfig.module_state_event"
