+++ subsys/caf/modules/CMakeLists.txt
@@ -16,7 +16,13 @@ zephyr_library_sources_ifdef(CONFIG_CAF_BUTTONS buttons.c)
 
 zephyr_library_sources_ifdef(CONFIG_CAF_CLICK_DETECTOR click_detector.c)
 
+zephyr_library_sources_ifdef(CONFIG_CAF_FACTORY_RESET_REQUEST factory_reset_request.c)
+
 zephyr_library_sources_ifdef(CONFIG_CAF_LEDS leds.c)
+
+zephyr_library_sources_ifdef(CONFIG_CAF_NET_STATE_LTE net_state_lte.c)
+
+zephyr_library_sources_ifdef(CONFIG_CAF_NET_STATE_OT net_state_ot.c)
 
 zephyr_library_sources_ifdef(CONFIG_CAF_POWER_MANAGER power_manager.c)
