+++ subsys/caf/modules/Kconfig
@@ -10,6 +10,8 @@ rsource "Kconfig.ble_smp"
 rsource "Kconfig.ble_state"
 rsource "Kconfig.buttons"
 rsource "Kconfig.click_detector"
+rsource "Kconfig.factory_reset_request"
 rsource "Kconfig.leds"
+rsource "Kconfig.net_state"
 rsource "Kconfig.power_manager"
 rsource "Kconfig.sensor_sampler"
