+++ subsys/caf/modules/Kconfig.net_state
@@ -1,0 +1,52 @@
+#
+# Copyright (c) 2021 Nordic Semiconductor ASA
+#
+# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+#
+
+menuconfig CAF_NET_STATE
+	bool "Networking state processing"
+	depends on LTE_LINK_CONTROL || NET_L2_OPENTHREAD
+	select CAF_NET_STATE_EVENTS
+	help
+	  Enable a module that process and sends network state events.
+
+if CAF_NET_STATE
+
+config CAF_LOG_NET_STATE_WAITING
+	bool "Log waiting on connection"
+	depends on LOG
+	default y
+	help
+	  If enabled periodic message informing about connection in progress
+	  will be emitted.
+
+config CAF_LOG_NET_STATE_WAITING_PERIOD
+	int "Waiting on connection log period [s]"
+	depends on CAF_LOG_NET_STATE_WAITING
+	range 1 3600
+	default 5
+	help
+	  Time in seconds at which waiting on connection log is emitted.
+
+config CAF_NET_STATE_LTE
+	bool
+	default CAF_NET_STATE && LTE_LINK_CONTROL
+
+config CAF_NET_STATE_OT
+	bool
+	default CAF_NET_STATE && NET_L2_OPENTHREAD
+
+if CAF_NET_STATE_LTE
+module = CAF_NET_STATE_LTE
+module-str = Network state LTE
+source "subsys/logging/Kconfig.template.log_config"
+endif
+
+if CAF_NET_STATE_OT
+module = CAF_NET_STATE_OT
+module-str = Network state OpenThread
+source "subsys/logging/Kconfig.template.log_config"
+endif
+
+endif
